fix compile errors on clang

+++ include/engine/engine.h
@@ -31,7 +31,7 @@ namespace ENGINE_NS {
         private:
             friend class Engine;
             uint64_t m_log_idx{};
-            std::array<RwLock<Logger>, static_cast<std::uint8_t>(LogNamespaces::COUNT)> loggers_;
+            std::array<RwLock<Logger>, static_cast<std::size_t>(LogNamespaces::COUNT)> loggers_;
             bool is_log_open_ = false;
     };
